+++ sdk/avs/arm-avs/src/azureVMwareSolutionAPI.ts
@@ -84,12 +84,15 @@ export class AzureVMwareSolutionAPI extends coreClient.ServiceClient {
       credential: credentials
     };
 
-    const packageDetails = `azsdk-js-arm-avs/3.1.2`;
+    const packageDetails = `azsdk-js-arm-avs/3.1.3`;
     const userAgentPrefix =
       options.userAgentOptions && options.userAgentOptions.userAgentPrefix
         ? `${options.userAgentOptions.userAgentPrefix} ${packageDetails}`
         : `${packageDetails}`;
 
+    if (!options.credentialScopes) {
+      options.credentialScopes = ["https://management.azure.com/.default"];
+    }
     const optionsWithDefaults = {
       ...defaults,
       ...options,