+++ sdk/monitor/monitor-opentelemetry-exporter/test/internal/base.exporter.test.ts
@@ -39,10 +39,10 @@ describe("#AzureMonitorBaseExporter", () => {
 
   it("should pass options to persister", () => {
     const exporter = new TestExporter();
-    assert.ok(exporter["_options"].instrumentationKey);
+    assert.ok(exporter["_instrumentationKey"]);
     assert.strictEqual(
-      (exporter["_persister"] as FileSystemPersist)["_options"].instrumentationKey,
-      exporter["_options"].instrumentationKey
+      (exporter["_persister"] as FileSystemPersist)["_instrumentationKey"],
+      exporter["_instrumentationKey"]
     );
     assert.deepStrictEqual(
       (exporter["_persister"] as FileSystemPersist)["_options"],
@@ -66,7 +66,7 @@ describe("#AzureMonitorBaseExporter", () => {
         nock.cleanAll();
       });
 
-      it("should persist retriable failed telemetry", async () => {
+      it("should persist retriable failed telemetry 429", async () => {
         const exporter = new TestExporter();
         const response = failedBreezeResponse(1, 429);
         scope.reply(429, JSON.stringify(response));
@@ -79,7 +79,7 @@ describe("#AzureMonitorBaseExporter", () => {
         assert.deepStrictEqual(persistedEnvelopes[0], toObject(envelope));
       });
 
-      it("should persist retriable failed telemetry", async () => {
+      it("should persist retriable failed telemetry 500", async () => {
         const exporter = new TestExporter();
         const response = failedBreezeResponse(1, 500);
         scope.reply(500, JSON.stringify(response));
@@ -92,7 +92,7 @@ describe("#AzureMonitorBaseExporter", () => {
         assert.deepStrictEqual(persistedEnvelopes[0], toObject(envelope));
       });
 
-      it("should persist retriable failed telemetry", async () => {
+      it("should persist retriable failed  502", async () => {
         const exporter = new TestExporter();
         const response = failedBreezeResponse(1, 502);
         scope.reply(502, JSON.stringify(response));
@@ -105,7 +105,7 @@ describe("#AzureMonitorBaseExporter", () => {
         assert.deepStrictEqual(persistedEnvelopes[0], toObject(envelope));
       });
 
-      it("should persist retriable failed telemetry", async () => {
+      it("should persist retriable failed telemetry 503", async () => {
         const exporter = new TestExporter();
         const response = failedBreezeResponse(1, 503);
         scope.reply(503, JSON.stringify(response));
@@ -118,7 +118,7 @@ describe("#AzureMonitorBaseExporter", () => {
         assert.deepStrictEqual(persistedEnvelopes[0], toObject(envelope));
       });
 
-      it("should persist retriable failed telemetry", async () => {
+      it("should persist retriable failed telemetry 504", async () => {
         const exporter = new TestExporter();
         const response = failedBreezeResponse(1, 504);
         scope.reply(504, JSON.stringify(response));