+++ sdk/appplatform/arm-appplatform/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 2.1.0-beta.3 (Unreleased)
+## 2.1.0-beta.4 (Unreleased)
 
 ### Features Added
 
@@ -9,6 +9,12 @@
 ### Bugs Fixed
 
 ### Other Changes
+
+## 2.1.0-beta.3 (2022-10-08)
+
+**Bugs Fixed**
+
+  -  revert credential scopes
 
 ## 2.1.0-beta.2 (2022-09-30)