+++ sdk/connectedvmware/arm-connectedvmware/CHANGELOG.md
@@ -1,10 +1,15 @@
 # Release History
 
-## 1.0.0-beta.2 (2022-09-30)
+## 1.0.0-beta.2 (Unreleased)
 
-**Bugs Fixed**
+### Features Added
 
-  -  fix better user experience of credential scopes in government cloud
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
 ## 1.0.0-beta.1 (2022-08-15)
 
 The package of @azure/arm-connectedvmware is using our next generation design principles. To learn more, please refer to our documentation [Quick Start](https://aka.ms/js-track2-quickstart).